Update #30: Enemy actions overhaul
The way enemy actions and behaviours was being handled was quite messy and I didn't really like it, so I did a complete overhaul, and now everything is controlled from the "Enemy" script, that gained a lot more importance after this. Now creating new actions should be much easier. It keeps the modularity it had but now it allows for some complex multiaction enemy (useful for bosses).

The system is not fully tested and some enemies might be broken (for example, LongWorm and Hopper are broken, but they were the glitchiest before the change)